refactor(dia.CellView)!: early evaluation of calc attributes
